refactor(interaction): extract InteractionController and InteractionService
